requirements.txt
openai
python-dotenv
youtube-dl
